Move user files to user_api directory
